Avoid overflow when computing vehicle power
When a vehicle was being fast-forwarded over a period of months upon
return to the reality bubble it was possible for the accumulated solar
power to exceed billions of joules.  This led to integer overflow.

Perform intermediate calculation in 64-bit ints to avoid the problem.

+++ src/vehicle.cpp
@@ -1229,7 +1229,9 @@ int vehicle::part_epower_w( const int index ) const
 int vehicle::power_to_energy_bat( const int power_w, const time_duration &d ) const
 {
     // Integrate constant epower (watts) over time to get units of battery energy
-    int energy_j = power_w * to_seconds<int>( d );
+    // Thousands of watts over millions of seconds can happen, so 32-bit int
+    // insufficient.
+    int64_t energy_j = power_w * to_seconds<int64_t>( d );
     int energy_bat = energy_j / bat_energy_j;
     int sign = power_w >= 0 ? 1 : -1;
     // energy_bat remainder results in chance at additional charge/discharge
